Move stuff from ADCL -> ADCL_Module...
Getting closer to the ADC stuff by moving most of the work into the module files.

Hopefully I did not screw everything up!
